requirements.txt
dnspython==2.6.1
mysql-connector-python==9.0.0
pymongo==4.8.0
